requirements.txt
torch
torchvision
pandas
matplotlib
seaborn
Pillow
scikit-learn
timm
tqdm
streamlit
plotly
jupyter
ipykernel
